backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
httpx==0.26.0
pillow==10.2.0
geoalchemy2==0.14.4
python-dotenv==1.0.1
shapely==2.0.3
pgvector==0.2.5
psycopg[binary]==3.1.18
torch==2.2.0
torchvision==0.17.0
open_clip_torch==2.24.0
numpy==1.26.4
huggingface_hub==0.20.3
